requirements.txt
hyperliquid-python-sdk==0.22.0
python-telegram-bot==21.0.1
websockets==13.1
python-dotenv==1.0.1
